Refactor to remove System-Component parity
