--- architect/components/__init__.py
@@ -1,3 +1,0 @@
-"""Component library."""
-from .component import Component  # isort:skip
-from . import diffractors, foreoptics, lenses, masks, sensors
