--- tests/test_systems/test_satellites/test_CubeSat.py
@@ -1,27 +1,0 @@
-"""Satellite class tests."""
-# stdlib
-import logging
-
-# external
-import astropy.units as unit
-
-# project
-from architect.systems.satellites import CubeSat
-
-LOG = logging.getLogger(__name__)
-
-
-def test_init():
-    """Test initialization."""
-    satellite = CubeSat()
-    LOG.info(satellite)
-
-
-def test_get_volume():
-    """Test get_dimensions() method."""
-    satellite = CubeSat(units=3)
-
-    result = satellite.get_volume()
-    LOG.info(result)
-
-    assert result.decompose().unit == unit.m**3
